sit code review for expression and codegen

Created-by: liuyu 00809192
Author-id: 53630
MR-id: 13149764
Commit-by: liuyu
Merged-by: lizheng 00465000
E2E-issues: issue299;issue297;issue296;issue295;issue294;issue292
Description:
+ [x] I've added/changed the documentation
+ [ ] I've added test for the code
已验证日构建用例及2G数据集的大数据联盟10条，没有问题。

See merge request Kunpeng-Computing-DC/BigData/OmniOperatorJIT!316

+++ core/src/codegen/string_util.h
@@ -117,7 +117,7 @@ public:
             }
         }
 
-        *outLen = s.length();
+        *outLen = static_cast<int32_t>(s.length());
         auto ret = omniruntime::codegen::ArenaAllocatorMalloc(contextPtr, *outLen + 1);
         error_t res = memcpy_s(ret, *outLen + 1, s.c_str(), s.length());
         if (res != EOK) {
